requirements.txt
fastapi
uvicorn
kafka-python
asyncio
fal_client
aiohttp
requests
pytest-asyncio
pytest